% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw.union.plot.R
\name{degree.of.hexagon}
\alias{degree.of.hexagon}
\title{Determine 'degree' of hexagonal pack i.e. how many rings}
\usage{
degree.of.hexagon(i)
}
\description{
O O O
    . O O O 
   . O O O O
    . O O O 
     . O O

(15 points) has degree = 3 (three rings, three points on each side; 
outermost ring is not full, with dots marking the empty spaces)

The number of points in each ring (excluding the innermost) proceeds
as an arithmetic progression: 6, 12, 18...

The cumulative number of points is the sum of the progression
i.e. an arithmetic series: 6, 18, 30 ...

So, the ith point is the sum of the arithmetic progression, plus change,
where the number of terms is the number of rings less one

The sum of the first n terms in an arithmetic progression is given by:

                  n(a_1 + a_n)
          S_n =  --------------
                        2

where a_j is the value of the jth term.

So we rearrange and solve for n with the quadratic equation(with a few
trimmings to account for the innermost ring and 1-indexing)
}

