% Generated by roxygen2 (4.0.1): do not edit by hand
\name{draw.union.plot}
\alias{draw.union.plot}
\title{Draw a union plot}
\usage{
draw.union.plot(OTUTable, GroupFactor = "Sample", ColourFactor = "Phylum",
  Pointsize = 1)
}
\arguments{
\item{OTUTable}{an OTU table in tidy format, i.e. a data frame with at least an "OTU"
column as well as columns for the group and colour factors.}

\item{GroupFactor}{a string corresponding to the name of a factor column in OTUTable,
which determines the three groups that will form the three cardinal regions in the plot.
Defaults to "Sample".}

\item{ColourFactor}{a string corresponding to the name of a factor column in OTUTable,
which determines the colours of the points. ColourFactor is mandatory; if you don't want
the points to be coloured, you should be drawing a Venn diagram instead. Defaults to
"Phylum".}

\item{Pointsize}{number to be passed to geom_point() as the point size. Defaults to 1.}
}
\value{
Returns a ggplot2 grob containing the union plot, which can then be viewed (e.g. with
print()) or saved (e.g. with ggsave()) at leisure. Note that the placement of text labels
in the plot is unfortunately imprecise; you'll probably need to go in and move them in
an image editing suite.
}
\description{
A union plot is like a Venn diagram, in that it shows the overlap between three
groups. Unlike a Venn diagram, the number of overlapping units (in this case, OTUs)
is not indicated by a number but by a point drawn for each OTU, which can then be
coloured by a factor of interest e.g. Phylum. Also unlike a Venn diagram, the regions
are not indicated by overlapping circles but by hexagons and trapezoids; the points
are arranged on a hexagonal grid. This allows the regions to be resized according
to the number of shared OTUs.

There's probably already a name for this kind of plot, but I couldn't find it after
some cursory googling so just named it "union plot". Email me if you know the actual
name and I'll fix it.
}

