% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{OTUTable}
\alias{OTUTable}
\title{An example OTU table.}
\format{A data frame with 1,202 rows and 4 columns:
\describe{
  \item{Sample}{the sample}
  \item{OTU}{the OTU}
  \item{Family}{the OTU's taxonomic family}
  \item{Count}{the OTU's count}
}}
\usage{
OTUTable
}
\description{
An example OTU table.
}
\keyword{datasets}
