% Generated by roxygen2 (4.0.1): do not edit by hand
\name{place.points.hexagon}
\alias{place.points.hexagon}
\title{Return cartesian coordinates for i points in hexagonal pack layout}
\usage{
place.points.hexagon(i)
}
\description{
Return cartesian coordinates for i points in hexagonal pack layout
}

