% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw.union.plot.R
\name{rotate.odd.r}
\alias{rotate.odd.r}
\title{Rotate odd-r points}
\usage{
rotate.odd.r(Points, Rotations = 1)
}
\description{
Given a set of points on a hex grid in odd-r coordinates,
rotate 60º widdershins about (0, 0)
}

