% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw.union.plot.R
\name{place.points.hexagon}
\alias{place.points.hexagon}
\title{Return cartesian coordinates for i points in hexagonal pack layout}
\usage{
place.points.hexagon(i)
}
\description{
Points are placed in a spiral fashion, from centre outwards
}

