% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw.union.plot.R
\name{place.points.trapezoid}
\alias{place.points.trapezoid}
\title{Place points in a trapezoid}
\usage{
place.points.trapezoid(n, p)
}
\description{
Return coordinates for n points, in trapezoidal layout with p points
in bottom row.

Coordinate system is hex, odd-r.
}

